working build config
