+++ package.json
@@ -21,8 +21,7 @@
   },
   "scripts": {
     "start": "webpack serve --mode development --config ./config/webpack.dev.js",
-    "build": "webpack build",
-    "test": "jest test"
+    "build": "webpack build --mode development --config ./config/webpack.production.js"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -59,6 +58,7 @@
     "webpack": "^5.99.5",
     "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^5.2.1",
+    "webpack-merge": "^6.0.1",
     "yarn": "^1.22.4"
   },
   "volta": {
